warning fix

+++ src/visage/animatrix.h
@@ -16,7 +16,7 @@ public:
   {
     if (loop == 0) // non-looping animatrices are always active
       return true;
-    uint32_t progress = tickCount() % loop;
+    uint64_t progress = tickCount() % loop;
     return progress >= start && progress < end;
   }
   inline bool isVisible() const { return visible; }
